refactor: change folder name utils to helpers
